multi_agent/requirements.txt
google-generativeai
anthropic
requests
